+++ crates/rayexec_execution/src/functions/scalar/arith/add.rs
@@ -211,3 +211,32 @@ impl PlannedScalarFunction for AddImpl {
         })
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use rayexec_bullet::{
+        array::{Array, Int32Array},
+        datatype::DataType,
+    };
+
+    use crate::functions::scalar::ScalarFunction;
+
+    use super::*;
+
+    #[test]
+    fn add_i32() {
+        let a = Arc::new(Array::Int32(Int32Array::from_iter([1, 2, 3])));
+        let b = Arc::new(Array::Int32(Int32Array::from_iter([4, 5, 6])));
+
+        let specialized = Add
+            .plan_from_datatypes(&[DataType::Int32, DataType::Int32])
+            .unwrap();
+
+        let out = specialized.execute(&[&a, &b]).unwrap();
+        let expected = Array::Int32(Int32Array::from_iter([5, 7, 9]));
+
+        assert_eq!(expected, out);
+    }
+}
