+++ crates/rayexec_execution/src/functions/scalar/arith/mod.rs
@@ -12,92 +12,3 @@ pub use mul::*;
 
 mod rem;
 pub use rem::*;
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use rayexec_bullet::{
-        array::{Array, Int32Array},
-        datatype::DataType,
-    };
-
-    use crate::functions::scalar::ScalarFunction;
-
-    use super::*;
-
-    #[test]
-    fn add_i32() {
-        let a = Arc::new(Array::Int32(Int32Array::from_iter([1, 2, 3])));
-        let b = Arc::new(Array::Int32(Int32Array::from_iter([4, 5, 6])));
-
-        let specialized = Add
-            .plan_from_datatypes(&[DataType::Int32, DataType::Int32])
-            .unwrap();
-
-        let out = specialized.execute(&[&a, &b]).unwrap();
-        let expected = Array::Int32(Int32Array::from_iter([5, 7, 9]));
-
-        assert_eq!(expected, out);
-    }
-
-    #[test]
-    fn sub_i32() {
-        let a = Arc::new(Array::Int32(Int32Array::from_iter([4, 5, 6])));
-        let b = Arc::new(Array::Int32(Int32Array::from_iter([1, 2, 3])));
-
-        let specialized = Sub
-            .plan_from_datatypes(&[DataType::Int32, DataType::Int32])
-            .unwrap();
-
-        let out = specialized.execute(&[&a, &b]).unwrap();
-        let expected = Array::Int32(Int32Array::from_iter([3, 3, 3]));
-
-        assert_eq!(expected, out);
-    }
-
-    #[test]
-    fn div_i32() {
-        let a = Arc::new(Array::Int32(Int32Array::from_iter([4, 5, 6])));
-        let b = Arc::new(Array::Int32(Int32Array::from_iter([1, 2, 3])));
-
-        let specialized = Div
-            .plan_from_datatypes(&[DataType::Int32, DataType::Int32])
-            .unwrap();
-
-        let out = specialized.execute(&[&a, &b]).unwrap();
-        let expected = Array::Int32(Int32Array::from_iter([4, 2, 2]));
-
-        assert_eq!(expected, out);
-    }
-
-    #[test]
-    fn rem_i32() {
-        let a = Arc::new(Array::Int32(Int32Array::from_iter([4, 5, 6])));
-        let b = Arc::new(Array::Int32(Int32Array::from_iter([1, 2, 3])));
-
-        let specialized = Rem
-            .plan_from_datatypes(&[DataType::Int32, DataType::Int32])
-            .unwrap();
-
-        let out = specialized.execute(&[&a, &b]).unwrap();
-        let expected = Array::Int32(Int32Array::from_iter([0, 1, 0]));
-
-        assert_eq!(expected, out);
-    }
-
-    #[test]
-    fn mul_i32() {
-        let a = Arc::new(Array::Int32(Int32Array::from_iter([4, 5, 6])));
-        let b = Arc::new(Array::Int32(Int32Array::from_iter([1, 2, 3])));
-
-        let specialized = Mul
-            .plan_from_datatypes(&[DataType::Int32, DataType::Int32])
-            .unwrap();
-
-        let out = specialized.execute(&[&a, &b]).unwrap();
-        let expected = Array::Int32(Int32Array::from_iter([4, 10, 18]));
-
-        assert_eq!(expected, out);
-    }
-}
